MULE-7443: Move SimpleLoggingTable to core so that it can be reused by other components without depending on launcher
