+++ core/src/main/java/org/mule/util/SimpleLoggingTable.java
@@ -4,9 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.module.launcher;
-
-import org.mule.util.StringUtils;
+package org.mule.util;
 
 import java.util.LinkedList;
 import java.util.List;
